Removed Prefabs, added new script.
A student at CSULB informed me that uploading prefabs and non-programming elements to GitHub from Unity can be a bother. Due to that, I decided to remove prefabs until I can come up with a better solution. In the meantime, I created a new script that allows for game over utilization, involving changing screens. Soon, I will make it so the game over text appears ontop of the grid and tetriminos. This was more of an experiment :)