Difficulty selection + last update from  the tutorial.
I finished the tetris tutorial!
That means everything from here on out is gonna be 100% from my brain.
I think I might need to make the game 3D though, since there is quite a bit I wanted to do but might be prevented by the 2d environment. More updates soon.

This is not the end